# swagger/documentations/authRoutesSwagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "127.0.0.1",
  "basePath": "/auth",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "post": {
        "tags": [
          "vitality"
        ],
        "description": "health check",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "description": "User Login",
        "parameters": [
          {
            "name": "credentials",
            "in": "body",
            "required": false,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "rni@sains.com.my"
                },
                "password": {
                  "type": "string",
                  "example": "g0g0Powerr@nger"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "description": "User Log Out",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/session-check": {
      "get": {
        "tags": [
          "auth"
        ],
        "description": "Checking session data stored in Redis and to see if user have a valid session",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}